Moved stuff to API-layer instead of direct access to backend from frontend
